+++ CHANGELOG.rst
@@ -1,7 +1,8 @@
 1.1.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Be able to use `iter_data` with custom uri
+  [vangheem]
 
 
 1.1.0 (2018-03-19)